Include on-disk view files; hide upload URL.
This always includes the on-disk view files; if they aren't needed in the
downstream `medic-os` build, they'll just be `rm`-ed before packaging starts.

Finally, the upload URL might have a username and password (it does currently),
which would show up in the publicly-accessible Travis log if `make` echoes it.
Don't echo it.