resolve_pending transition should be disabled on final app settings.
We need to include resolve_pending transition during messages data load
so that test messages are not pending on first run.  So we enable it
temporarily during the messages load but make sure it's not enabled on
final ddoc.

Added a third parameter to load_settings script that takes an optional
js module to munge the settings before saving to the ddoc.